requirements.txt
numpy
scipy
pyyaml
matplotlib
networkx
